requirements.txt
streamlit==1.35.0
openai==1.25.0
pandas==2.2.0
python-dotenv==1.0.0
tqdm==4.66.1 